Making ivoa-cite work with TTH.
Note: I'm considering xspace too hard for TTH (I've tried, and it's not
straightforward) and using ~ (or whatever) not that much of a big deal.  Hence:
in contrast to the previous implementation, no automatic spacing is being done.
I'll ask TTH's author about xspace, though.

Document changes exclusively for abolishion of xspace.