Fix grammar of `UNION`, `EXCEPT` and `INTERSECT`.
A table name or join is no longer directly allowed as left or right
operand of a set operator (on the contrary to what SQL92 seems to allow ;
the SQL92 rules are not supported by the nowadays major DBMS).

Without parentheses, only a `SELECT` without `WITH`, `ORDER BY` and `OFFSET`
clauses is possible. Note that any set operations are still possible in such
limited `SELECT` query.

With parentheses, a complete `SELECT` query (so, with `WITH`, `ORDER BY`
and `OFFSET` clauses) is possible.

Set operations are still evaluated with the following precedence:

1. Expressions in parentheses
2. The `INTERSECT` operation
3. `EXCEPT` and `UNION` (from left to right)