Clarify that geometry types and functions are spherical
There are relics in the text of the (STC-1?) possibility of
using Cartesian coordinates alongside spherical coordinates
for ADQL geometries and data types.  Adjust the text to
clarify that geometries in ADQL are always spherical.

This includes making this explicit in the references to DALI
which is itself agnostic about coordinate system
(e.g. DALI 1.1 sec 3.3.5 says "Geometry values are two-dimensional;
although they are usually longitude and latitude values in spherical
coordinates this is specified in the coordinate metadata and not
in the values").