Change `\clearpage` and `\hline` usage.
Usage of `\clearpage` was not very clear. Although it seems to not be a very
clean LaTeX practice, it has been decided here to use it in only two cases:

1. to force the layout of a block (e.g. figure, verbatim, table) ; typically to
   force the VOArch figure to be along its description and not randomly in
   another subsection ; or to force the ADQL Query Syntax block to be entirely
   visible on a single page (instead of being split up)

2. to start each section (and only sections, not (sub)subsections) on a new page

Changing this `\clearpage` usage leads to remove a small paragraph inherited
from ADQL-2.0 and introducing the description of geometrical functions. It seems
to be useless considering the upper subsection is named "Geometrical functions".
Then, it makes sense to have subsubsections about geometrical functions.

About `\hline`, it was doubled on the beginning of several tables. The sideback
of this practice was to generate an empty row in the HTML version. The same for
an empty line just before the final `\hline` in several tables ; it generated
an empty row at the end of tables in the HTML version.