Pull BNF out of main document
The BNF is now in a separate file adql.bnf, imported using
\lstinputlisting.  This makes it more straightfoward to work
with the BNF as a separate item.
I also replaced the single tab character in the BNF with 8 spaces.

The \lstinputlisting is configured to make it look the same as
it used to in the \verbatim environment that was there before.

fixup pull