DOUBLE and PRECISION are SQL, not ADQL reserved words
DOUBLE and PRECISION are now entries in the list of SQL
reserved words not ADQL reserved words as included in both
the Section 2 text and the BNF, as they were in ADQL 2.0
(I believe they are both SQL92 reserved words).

The main reason for this change was to remove the item
"DOUBLE PRECISION" from the "CAST: function and datatypes"
sublist in the ADQL reserved keywords section 2.1.4 of the text,
on the grounds that an item with embedded whitespace
doesn't look right in a list of tokens.