Review coordinate limits (#99)
Update coordinate limits section. No more explicit coordinate limits (as it would need to give limits for all possible coordinate systems) but make a reference to DALI-1.1 and future versions.

There is going to be an erratum to DALI 1.1 specifying that the given coordinate limits in DALI are for equatorial systems. In next version of DALI (1.2) there will be limits for more coordinate systems.